Add deposit and balance functionality

+++ server/models/expense.go
@@ -3,6 +3,6 @@ package models
 type Expense struct {
 	Id                int     `json:"id" db:"id"`
 	Title             string  `json:"title" db:"title"`
-	Value             float32 `json:"value" db:"value"`
+	Value             float64 `json:"value" db:"value"`
 	ExpenseCategoryId int     `json:"expense_category_id" db:"expense_category_id"`
 }
