+++ server/models/expense_overview.go
@@ -6,6 +6,6 @@ type ExpenseOverview struct {
 	UserId        int             `json:"user_id" db:"user_id"`
 	CategoryId    int             `json:"category_id" db:"category_id"`
 	CategoryName  string          `json:"name" db:"category_name"`
-	CategoryTotal float32         `json:"total" db:"category_total"`
+	CategoryTotal float64         `json:"total" db:"category_total"`
 	ExpenseSum    sql.NullFloat64 `json:"expense_sum" db:"sum_expenses"`
 }
